+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Update to newer libswresample library
 * Reload lua scripts when modified by user (#519)
 * Debug errors show file and line
+* Add markers to frames in input editor (#548)
 
 ### Changed